+++ src/index.ts
@@ -18,7 +18,6 @@ export { createThemeTransition } from './createThemeTransition';
 export type {
   ThemeTransitionConfig,
   ThemeTransitionAPI,
-  SystemThemeMap,
   SetThemeOptions,
   ThemeDefinition,
   ThemeNames,